% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnca.R
\name{ggnca}
\alias{ggnca}
\title{ggplot for pharmacokinetic concentration-time curve}
\usage{
ggnca(concData, colSubj = "Subject", colTime = "Time", colConc = "conc")
}
\arguments{
\item{concData}{data}

\item{colSubj}{subject}

\item{colTime}{Time}

\item{colConc}{conc}
}
\description{
This draws pharmacokinetic concentration-time curve with ggplot2 packages.
}
\section{}{
NA
}

