% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansdcv.R
\name{meansdcv}
\alias{meansdcv}
\title{Combining mean, standard deviation and coefficient of variation}
\usage{
meansdcv(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This function combines mean, standard deviation and coefficient of variation to create reporting tables.
}
\examples{
library(dplyr)
tabNCA(Theoph, dose=500, concUnit="mg/L") \%>\%
  as.data.frame() \%>\%
  summarise_all(meansdcv)
}
